refactor: extract vector-only routes to src/routes/vector/ (#1071 phase 1.1)
Move every vector-only HTTP endpoint into a dedicated module so the vector
layer can later move behind a network proxy (VECTOR_URL, phase 1.2) without
disturbing FTS / hybrid search.

Routes (URLs unchanged — backward compatible):
  - GET /api/similar         (was src/routes/search/similar.ts)
  - GET /api/compare         (was src/routes/compare/)
  - GET /api/map             (was src/routes/search/map.ts)
  - GET /api/map3d           (was src/routes/search/map3d.ts)
  - GET /api/vector/stats    (new — same payload as /api/stats vector block)
  - GET /api/vector/health   (new — adapter liveness probe, 503 when down)

Handlers:
  - New src/server/vector-handlers.ts owns handleSimilar, handleMap,
    handleMap3d, handleVectorStats, handleVectorHealth.
  - src/server/handlers.ts keeps handleSearch (hybrid), handleList,
    handleReflect, handleStats, handleGraph, handleLearn, etc.
  - Inlined the private getVectorStore() pass-through in handleSearch
    (now calls ensureVectorStoreConnected directly).

Wiring:
  - src/server.ts: compareRoutes → vectorRoutes (vectorRoutes also covers
    /api/compare).
  - src/routes/health/stats.ts: imports handleVectorStats from the new
    vector-handlers module.
  - src/routes/search/{index,model}.ts: drop the moved endpoints + their
    TypeBox schemas.

Verified:
  - bun tsc --noEmit clean.
  - bun test tests/http/{compare,knowledge}.test.ts → 32/32 pass (the two
    HTTP suites that exercise these endpoints).
  - bun test src/tools/__tests__/search.test.ts → 31/31 pass.
  - Full bun test has 80 pre-existing failures (SQLite disk-I/O lock
    races on menu_items + adapter env-bound tests); none touch the
    refactored code paths.
