+++ src/routes/search/index.ts
@@ -1,19 +1,16 @@
 /**
- * Search Routes (Elysia) — composes /api/{search,reflect,similar,map,map3d,list}.
+ * Search Routes (Elysia) — composes /api/{search,reflect,list}.
+ *
+ * Vector-only endpoints (similar, map, map3d, compare) live in
+ * src/routes/vector/ since #1071 phase 1.1.
  */
 
 import { Elysia } from 'elysia';
 import { searchEndpoint } from './search.ts';
 import { reflectEndpoint } from './reflect.ts';
-import { similarEndpoint } from './similar.ts';
-import { mapEndpoint } from './map.ts';
-import { map3dEndpoint } from './map3d.ts';
 import { listEndpoint } from './list.ts';
 
 export const searchRoutes = new Elysia({ prefix: '/api' })
   .use(searchEndpoint)
   .use(reflectEndpoint)
-  .use(similarEndpoint)
-  .use(mapEndpoint)
-  .use(map3dEndpoint)
   .use(listEndpoint);
