+++ src/routes/search/model.ts
@@ -1,5 +1,8 @@
 /**
- * TypeBox schemas for search routes.
+ * TypeBox schemas for search routes (FTS / hybrid).
+ *
+ * Vector-only schemas (SimilarQuery, Map3dQuery, CompareQuery) live in
+ * src/routes/vector/model.ts.
  */
 
 import { t } from 'elysia';
@@ -12,16 +15,6 @@ export const SearchQuery = t.Object({
   mode: t.Optional(t.String()),
   project: t.Optional(t.String()),
   cwd: t.Optional(t.String()),
-  model: t.Optional(t.String()),
-});
-
-export const SimilarQuery = t.Object({
-  id: t.Optional(t.String()),
-  limit: t.Optional(t.String()),
-  model: t.Optional(t.String()),
-});
-
-export const Map3dQuery = t.Object({
   model: t.Optional(t.String()),
 });
